changed /bin/server to use configfiles
